Fixed errors, improved graphics
The plots now go from 0 to L with the the major grid lines equal to the
delta x gridblocks. Added a well identifier on the graph for each well.
Fixed issue of grids not lining up to wells, block 1 not plotting, and
tweaked some background equations. Still can't figure out if the
constant pressure case going to <0 is good or bad.